Changed IsNationalDexEnabled Check to GameClear Check
Every important check for the national dex has been replaced by a game clear check. These are:
NPC blocking Elite Four if National Dex is acquired. This has been changed to her blocking it if the Game is Cleared but she will move once the Sevii Islands are completed.
Celio: If the player has cleared the game, he'll open the path to Islands 4+ instead of if the player has the national dex

We're good to add the national dex into the early game now.